fix error and add page

+++ telegram_bot/auth/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Float, ForeignKey, DateTime, Boolean
+from sqlalchemy import Integer, String, Float, ForeignKey, DateTime, Boolean, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from src.db.database import Base
 from datetime import datetime
@@ -9,6 +9,6 @@ class User_tg(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("user.id", ondelete="CASCADE"))
     user: Mapped["User"] = relationship(back_populates="tg_user")
-    tg_id: Mapped[int] = mapped_column(Integer)
+    tg_id: Mapped[int] = mapped_column(BigInteger)
     login: Mapped[bool] = mapped_column(Boolean, default=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
